Add check that values input into evaluate are valid
