+++ src/causalprog/graph/node/base.py
@@ -168,3 +168,19 @@ class Node(Labelled):
             List of labels of parent nodes
 
         """
+
+    @abstractmethod
+    def is_valid_value(self, value: float | npt.NDArray[float]) -> bool:
+        """Check if a value is valid for this node."""
+
+    def assert_is_valid_value(self, value: float | npt.NDArray[float]) -> None:
+        """Check if a value is valid for this node."""
+        if not self.is_valid_value(value):
+            msg = (
+                f"Invalid value for {self.__class__.__name__}: "
+                f"{self.label} cannot be {value}"
+            )
+            raise ValueError(msg)
+        if self.shape != (value.shape if hasattr(value, "shape") else ()):
+            msg = f"Invalid value for node: {self.label}"
+            raise ValueError(msg)
