+++ src/causalprog/graph/node/constant.py
@@ -61,3 +61,7 @@ class ConstantNode(Node):
     @property
     def parents(self) -> list[str]:
         return []
+
+    @override
+    def is_valid_value(self, value: float | npt.NDArray[float]) -> bool:
+        return True
