+++ src/causalprog/graph/node/data.py
@@ -67,3 +67,7 @@ class DataNode(Node):
     @property
     def parents(self) -> list[str]:
         return []
+
+    @override
+    def is_valid_value(self, value: float | npt.NDArray[float]) -> bool:
+        return True
